Get Employee API
First I made a metods in service interface which gets employee by id then I made an implementation in class which creates object that finds the id in repository if it could not finds it makes an exception. I made a exception class to give a message that says for not finding any id. After that I made a rest api for get mapping.